fix(sat): arena stride-desync zeroed live clause headers (latent, exposed by f32); activity f64->f32, 12B header
Root cause (caught by gdb hardware watchpoint on the exact zeroed
field; .#debug shell from b8f94ad): scale_live_activities walked the
arena recomputing each slot stride as slot_size(current_len), but a
clause's physical slot size is fixed at allocation while shrink later
lowers len. After a 4->3-literal shrink the walk's stride (32->24)
diverged, landed mid-slot, and treated a bogus offset as live; the
rescale's f32 read-modify-write 8 bytes past it read a real clause's
len field (raw 4) as the denormal 5.6e-45 and multiplied by 1e-20,
underflowing to an exact zero bit pattern. len=0 reached propagation;
crn_11_99_u answered sat (false).

Latent since the arena commit (89cacf4), not introduced by f32: the
identical walk existed under the f64 header, unreachable at
verification scale because the rescale trigger sat at increment>1e100
(~230k conflicts) above every corpus cap. Lowering the bound to
MiniSat's 1e20 - mandatory for f32 activities, whose range ends at
3.4e38 (the 1e100 policy would saturate f32 to inf and collapse the
reduce ordering) - moved the first rescale to ~46k conflicts and
exposed it inside the regression suite.

Fix: activity scaling is refs-driven (the refs table is the
authoritative slot list, exact under shrink by construction); the
arena gains per-ref scale_activity and the stride-walking
implementations are deleted. Regression: alloc -> shrink across the
32->24 stride boundary -> scale by 1e-20 -> every clause keeps its
true length.

Also in this change: activity f64->f32 shrinks the header 16->12
bytes (Lit gains #[repr(transparent)]); pinned geometry tests: a
3-literal slot is 24B so two ternary clauses share one cache line
(48B used), and a 5-literal clause is exactly half a line (32B; 4 was
the f64-header ceiling). Perf at 300k fixed conflicts on stable-300:
neutral (18.6s -> 18.6s interleaved x3) - the win is density and the
pinned properties.

Verification: 852 oxiz-sat + 10036 workspace tests, clippy/fmt/rustdoc
clean, 6k differential fuzz iterations, z3 parity 168 files 0 wrong.

+++ oxiz-sat/src/solver/conflict.rs
@@ -269,9 +269,11 @@ impl Solver {
                 if self.clauses.get(reason_clause).is_some_and(|c| c.lbd <= 2) {
                     self.clauses.promote_to_core(reason_clause);
                 }
-                // Bump clause activity (MapleSAT-style)
+                // Bump clause activity (MapleSAT-style). The increment
+                // schedule stays f64 (precision of the growth/decay cycle);
+                // the accumulated activity itself is f32 in the arena.
                 self.clauses
-                    .bump_activity(reason_clause, self.clause_bump_increment);
+                    .bump_activity(reason_clause, self.clause_bump_increment as f32);
             }
 
             let Some(clause) = self.clauses.get(reason_clause) else {
